requirements.txt
streamlit==0.73.1
joblib==0.14.1
pandas==1.0.1
Pillow==8.1.0
